# async batch verification