stacked on #9132 

Its difficult to give a solid prediction of how this will perform in
production, but here is some data on the raw CPU time savings:

With async batch verification:

    $ time RUST_LOG=off SIM_STRESS_TEST_QPS=30  ../../target/simulator/deps/simtest-3e38176b1c2a7e11 test_simulated_load_basic

    running 1 test
    test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 6 filtered out; finished in 15.82s

    RUST_LOG=off SIM_STRESS_TEST_QPS=30  test_simulated_load_basic  14.53s user 0.97s system 97% cpu 15.851 total
                                                                    ^^^^^^^^^^^

Without batch verification:

    ; time RUST_LOG=off SIM_STRESS_TEST_QPS=30  ../../target/simulator/deps/simtest-3e38176b1c2a7e11 test_simulated_load_basic

    running 1 test
    test result: ok. 1 passed; 0 failed; 0 ignored; 0 measured; 6 filtered out; finished in 17.91s

    RUST_LOG=off SIM_STRESS_TEST_QPS=30  test_simulated_load_basic  16.72s user 1.06s system 98% cpu 18.065 total
                                                                    ^^^^^^^^^^^

These numbers are corroborated by profiles taken with xcode, which indicated
a drop from 55 Gigacycles down to 48Gc, about a 15% improvement.

Lastly, the async benchmarks show that 1600 certs per second per core is achievable
with the async interface. The single-threaded synchronous benchmarks top out at a
bit over 2000 certs per second. We may be able to improve the async interface a bit
more with a lower-contention locking method.