@mystenmark 2023-03-12
> Is there a simple way to define the parameters (buffer len 8 & timeout of 10) as locally configurable at runtime?

I started doing this and it got ugly really fast. We don't have a good way to plumb config values through our whole system. I would like to add this but it will need to be in another PR.